+++ src/components/Button/index.tsx
@@ -5,10 +5,9 @@ import type { DefaultTheme, FlattenInterpolation, ThemeProps } from 'styled-comp
 import styled from 'styled-components';
 import { Spinner } from '#src/components/Spinner';
 import { appearanceMixin } from './appearanceMixin';
-import { dimensionMixin } from './dimensionMixin';
+import { ButtonIconContainer, dimensionMixin } from './dimensionMixin';
 import { mediumGroupBorderRadius } from '#src/components/themes/borderRadius';
 import { skeletonAnimationMixin } from '#src/components/skeleton/animation';
-import { IconContainer } from '#src/components/TextButton/commonMixin';
 
 export type { ButtonAppearance } from './types';
 export * from './dimensionMixin';
@@ -34,10 +33,22 @@ export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   /** Состояние скелетона */
   skeleton?: boolean;
 
-  /** Иконка кнопки */
+  /**
+   * @deprecated Используйте iconStart или iconEnd
+   * Иконка кнопки
+   */
   icon?: ReactNode;
 
-  /** Распооложение иконки кнопки */
+  /** Иконка перед текстом кнопки */
+  iconStart?: ReactNode;
+
+  /** Иконка после текста кнопки */
+  iconEnd?: ReactNode;
+
+  /**
+   * @deprecated Используйте iconStart или iconEnd
+   * Расположение иконки кнопки
+   */
   iconPlace?: IconPlace;
 
   /** Позволяет добавлять миксин для кнопок, созданный с помощью styled css  */
@@ -48,6 +59,7 @@ const ButtonContent = styled.div<{ dimension?: Dimension; $loading?: boolean }>`
   vertical-align: top;
 
   display: inline-flex;
+  gap: 8px;
   flex-direction: row;
   overflow: hidden;
   flex-wrap: nowrap;
@@ -59,10 +71,6 @@ const ButtonContent = styled.div<{ dimension?: Dimension; $loading?: boolean }>`
     display: inline-block;
     flex: 0 1 auto;
     white-space: nowrap;
-  }
-
-  > *:not(:first-child) {
-    margin-left: 8px;
   }
 
   & > svg {
@@ -87,6 +95,11 @@ const StyledSpinner = styled(Spinner)`
   transform: translate(-50%, -50%);
 `;
 
+const AdditionalPadding = styled.div`
+  display: inline-block;
+  width: 2px;
+`;
+
 export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   (
     {
@@ -95,6 +108,8 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
       type = 'button',
       loading = false,
       skeleton = false,
+      iconStart,
+      iconEnd,
       icon,
       iconPlace = 'left',
       children,
@@ -105,8 +120,8 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   ) => {
     const spinnerDimension = dimension === 's' ? 'ms' : 'm';
     const spinnerInverse = appearance !== 'secondary' && appearance !== 'ghost';
-    const hasIconLeft = !!icon && iconPlace === 'left';
-    const hasIconRight = !!icon && iconPlace === 'right';
+    const hasIconStart = !!iconStart || (!!icon && iconPlace === 'left');
+    const hasIconEnd = !!iconEnd || (!!icon && iconPlace === 'right');
 
     return (
       <StyledButton
@@ -116,19 +131,19 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
         type={type}
         $loading={loading}
         skeleton={skeleton}
-        hasIconLeft={hasIconLeft}
-        hasIconRight={hasIconRight}
         buttonCssMixin={buttonCssMixin}
         {...props}
       >
         {loading && <StyledSpinner dimension={spinnerDimension} inverse={spinnerInverse} />}
+        {!loading && !props.displayAsSquare && !hasIconStart && <AdditionalPadding />}
         <ButtonContent>
-          {hasIconLeft && <IconContainer>{icon}</IconContainer>}
+          {hasIconStart ? <ButtonIconContainer>{iconStart || icon}</ButtonIconContainer> : null}
           {React.Children.toArray(children).map((child, index) =>
             typeof child === 'string' ? <div key={child + index}>{child}</div> : child,
           )}
-          {hasIconRight && <IconContainer>{icon}</IconContainer>}
+          {hasIconEnd ? <ButtonIconContainer>{iconEnd || icon}</ButtonIconContainer> : null}
         </ButtonContent>
+        {!loading && !props.displayAsSquare && !hasIconEnd && <AdditionalPadding />}
       </StyledButton>
     );
   },
@@ -146,10 +161,11 @@ const StyledButton = styled.button.attrs<
   position: relative;
   box-sizing: border-box;
   display: inline-block;
+  white-space: nowrap;
   border: none;
   border-radius: ${(p) => (p.skeleton ? 0 : mediumGroupBorderRadius(p.theme.shape))};
   appearance: none;
-  vertical-align: center;
+  vertical-align: middle;
   pointer-events: ${(p) => (p.$loading || p.disabled || p.skeleton ? 'none' : 'all')};
 
   ${appearanceMixin};